Pass context to map
Parsers may want to attach contextual information, such as line number
and position, to the AST nodes they are creating while parsing. One way
to create such AST nodes is by transforming rule results with `map`.
However, the current `map` implementation only receives the results
generated by the rule, so there is no way to grab the current context.

This PR attempts to solve this problem by passing the `ctx` object as
a second argument to map. This seems to maintain backwards compatibility
because Procs can ignore this parameter. However, I am not familiar with
the codebase and do not understand if there are other implications or
other ways to achieve the desired resul.